Move period units to periods.py
