+++ openfisca_core/columns.py
@@ -19,11 +19,6 @@ def N_(message):
 
 year_or_month_or_day_re = re.compile(ur'(18|19|20)\d{2}(-(0?[1-9]|1[0-2])(-([0-2]?\d|3[0-1]))?)?$')
 
-# Possible values for period_unit
-MONTH = u'month'
-YEAR = u'year'
-PERMANENT = u'permanent'
-
 
 # Base Column
