+++ openfisca_core/tests/test_extra_params.py
@@ -2,7 +2,8 @@
 
 
 from openfisca_core import periods
-from openfisca_core.columns import IntCol, BoolCol, MONTH
+from openfisca_core.columns import IntCol, BoolCol
+from openfisca_core.periods import MONTH
 from openfisca_core.variables import Variable
 from openfisca_core.tests import dummy_country
 from openfisca_core.tests.dummy_country.entities import Individu
