+++ openfisca_core/tests/test_opt_out_cache.py
@@ -3,7 +3,8 @@
 
 from openfisca_core.tests import dummy_country
 from openfisca_core.variables import Variable
-from openfisca_core.columns import IntCol, MONTH
+from openfisca_core.columns import IntCol
+from openfisca_core.periods import MONTH
 from openfisca_core.tests.dummy_country.entities import Individu
